requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==2.6.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
